Update io.itch.itch.yaml
removed the `--filesystem` line for `xdg-data/icons:ro` as the build-bot complained about it being unnecessary. 

Not sure if that was only a necessity on NixOS
$HOME/.local/share/icons should exist on all systems and be populated with the chosen themes cursor-icon-files. 

The need to explicitly state it seems to be a wayland peculiarity anyway as the cursor is actually grabbed by wayland and xorg both from the same files. There are no wayland specific cursor files (yet).